8.1, 8.2 - deleting acc/cat with nonzero balance

+++ api/src/main/java/com/nathangawith/umkc/services/SettingsService.java
@@ -44,12 +44,16 @@ public class SettingsService implements ISettingsService {
 	@Override
 	public boolean removeAccount(int userID, int id, boolean showWarning) throws Exception {
 		boolean isAccountUsed = mSettingsRepository.isAccountUsedInTransactions(userID, id);
+		DBAccount account = mSettingsRepository.selectAccount(userID, id);
+		boolean isZeroBalance = account == null ? true : account.Total == 0;
 		if (isAccountUsed && showWarning) throw new Exception(Messages.ACCOUNT_USED_IN_TRANSACTIONS);
-		if (isAccountUsed) {
-			return mSettingsRepository.updateAccountDisable(userID, id);
-		} else {
-			return mSettingsRepository.deleteAccount(userID, id);
-		}
+		if (isZeroBalance) {
+			if (isAccountUsed) {
+				return mSettingsRepository.updateAccountDisable(userID, id);
+			} else {
+				return mSettingsRepository.deleteAccount(userID, id);
+			}
+		} else throw new Exception(Messages.ACCOUNT_HAS_NONZERO_BALANCE);
 	}
 
 	@Override
@@ -77,23 +81,38 @@ public class SettingsService implements ISettingsService {
 	public Collection<DBCategory> getCategories(int userID, String categoryType) throws Exception {
 		return mSettingsRepository.selectCategories(userID, categoryType);
 	}
+	
+	private DBCategory getCategoryWithUnknownType(int userID, Integer id, String description) {
+		DBCategory incomeCategory = null;
+		DBCategory expenseCategory = null;
+		if (description == null) {
+			incomeCategory = mSettingsRepository.selectCategory(userID, Constants.INCOME, id);
+			expenseCategory = mSettingsRepository.selectCategory(userID, Constants.EXPENSE, id);
+		} else if (id == null) {
+			incomeCategory = mSettingsRepository.selectCategory(userID, Constants.INCOME, description);
+			expenseCategory = mSettingsRepository.selectCategory(userID, Constants.EXPENSE, description);
+		}
+		return incomeCategory != null ? incomeCategory : expenseCategory;
+	}
 
 	@Override
 	public boolean removeCategory(int userID, int id, boolean showWarning) throws Exception {
 		boolean isCategoryUsed = mSettingsRepository.isCategoryUsedInTransactions(userID, id);
+		DBCategory category = this.getCategoryWithUnknownType(userID, id, null);
+		boolean isZeroBalance = category == null ? true : category.Total == 0;
 		if (isCategoryUsed && showWarning) throw new Exception(Messages.CATEGORY_USED_IN_TRANSACTIONS);
-		if (isCategoryUsed) {
-			return mSettingsRepository.updateCategoryDisable(userID, id);
-		} else {
-			return mSettingsRepository.deleteCategory(userID, id);
-		}
+		if (isZeroBalance) {
+			if (isCategoryUsed) {
+				return mSettingsRepository.updateCategoryDisable(userID, id);
+			} else {
+				return mSettingsRepository.deleteCategory(userID, id);
+			}
+		} else throw new Exception(Messages.CATEGORY_HAS_NONZERO_BALANCE);
 	}
 
 	@Override
 	public boolean editCategory(int userID, int id, String description) throws Exception {
-		DBCategory incomeCategory = mSettingsRepository.selectCategory(userID, Constants.INCOME, description);
-		DBCategory expenseCategory = mSettingsRepository.selectCategory(userID, Constants.EXPENSE, description);
-		DBCategory category = incomeCategory != null ? incomeCategory : expenseCategory;
+		DBCategory category = this.getCategoryWithUnknownType(userID, null, description);
 		if (category != null) {
 			if (category.Enabled == 1) throw new Exception(Messages.CATEGORY_ALREADY_EXISTS);
 			else throw new Exception(Messages.CATEGORY_ALREADY_DISABLED);
